SportDevils upgrade for AceStream and Sopcast
Here an upgrade for SportDevils to support acestream and sopcast steams. This is in particular working for livetv.sx. 
The upgrade is based on plexus plugin that must be installed on the Kodi. 
No new module has been imported to maintv.py. No platform specific instruction have been added. 
Tested on firestick (android) with two kodi 17